requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
deepl==1.11.0
ffmpeg-python==0.2.0
filelock==3.9.0
future==0.18.2
huggingface-hub==0.11.1
idna==3.4
more-itertools==9.0.0
numpy==1.24.1
omegaconf==2.3.0
packaging==22.0
pycparser==2.21
python-dotenv==0.21.0
python-http-client==3.3.7
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
silero==0.4.1
soundfile==0.11.0
tokenizers==0.13.2
torch==1.13.1
torchaudio==0.13.1
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
urllib3==1.26.13
whisper==1.0
